% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_std_error.R
\name{est_se}
\alias{est_se}
\title{Extract standard error}
\usage{
est_se(n, ci = 0.95, ul, ll, diff, p = NA)
}
\arguments{
\item{n}{number of patients}

\item{ci}{It's the percentile that is used for finding the quantile function from a t-distribution with df degrees of freedom. Default is CI 95% which correspond to perc .975 (Others ci may be used).  The quantile will be differently obtained if p_value is provided.}

\item{diff}{The difference}

\item{upper_limit}{upper limit of confidence interval}

\item{lower_limit}{lower limit of confidence interval}

\item{p_value}{It is assumed that p is equal to the upper bound, difference is required.}
}
\value{
SE_diff
}
\description{
Estimate the standard error
}
\author{
Carlos S Traynor
}
\keyword{se}
