% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_meta.R
\name{ic_meta}
\alias{ic_meta}
\alias{ic_meta.data.frame}
\title{Bayesian Hierarchical Model for Meta-Analysis}
\usage{
ic_meta(object, ...)

\method{ic_meta}{data.frame}(object, transform = no_trans,
  hetero_var = TRUE, ...)
}
\arguments{
\item{object}{A data frame containing the 'study' column, the 'yi' observed effect column ,  and at least one of the columns 'vi' variance or ; 'sei' standard error.}

\item{...}{Additonal arguments to pass to [brms::brm()]
such as `set_prior`, `seed`, `warmup`, etc.}

\item{transform}{An named list of transformation and inverse
transformation fuctions. See [logit_trans()] as an example.}

\item{hetero_var}{A logical; if `TRUE`, then different
variances are estimated for each model group. Otherwise, the
same variance is used for each group akinf to the fixed effect model in the
frequentist literature.}
}
\value{
An object of class `ic_meta`.
}
\description{
Bayesian hierarchical model used here to answer the question does the
compiled literature provide evidence for the effectiveness of the
treatment.
}
\details{
These functions can be used to perform a Meta-analysis via
Bayesian multilevel modeling.

By default, a generalized linear model with Gaussian error and
 an identity link is fit to the data.
}
\examples{



# Example objects from the "Getting Started" vignette at
#  https://topepo.github.io/tidyposterior/articles/Getting_Started.html

file <- system.file("examples", "roc_model.RData", package = "tidyposterior")
load(file)

roc_model

# Summary method shows the underlying `stan` model
summary(roc_model)
}
