% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.R
\name{est_se}
\alias{est_se}
\title{Estimate the standard error}
\usage{
est_se(n, level = 0.95, ul, ll, diff, p = NA)
}
\arguments{
\item{n}{number of patients}

\item{level}{It's the level of the ci which is used for finding the quantile function
from a t-distribution with df degrees of freedom. Default is CI 95% which correspond
 to perc .975 (Others ci may be used).  The quantile will be differently obtained
 if p_value is provided.}

\item{ul}{upper limit of confidence interval}

\item{ll}{lower limit of confidence interval}

\item{diff}{The difference}

\item{p}{It is assumed that p is equal to the upper bound, difference is required.}
}
\value{
est_se
}
\description{
Extract relevant features from a glmnet cox fit object.
}
\author{
Carlos S Traynor
}
\keyword{se}
